integrations/hongshan/hongshan-backend/requirements.txt
# 红山量化交易平台 - 后端依赖

# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 行情数据
akshare==1.13.28
pandas==2.1.4
numpy==1.26.3

# 密码加密
passlib[bcrypt]==1.7.4

# HTTP 请求
requests==2.31.0

# 缓存与 JWT（与 app 代码一致）
redis==5.0.1
PyJWT==2.8.0

# 日志
python-json-logger==2.0.7

# 开发工具
python-dotenv==1.0.0
